pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-cloud_devops_env"
version = "0.1.0"
description = "Cloud Devops Env environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "pydantic>=2.0.0",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
# Server entry point - enables running via: uv run --project . server
# or: python -m cloud_devops_env.server.app
server = "cloud_devops_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["cloud_devops_env", "cloud_devops_env.server"]
package-dir = { "cloud_devops_env" = ".", "cloud_devops_env.server" = "server" }